#185 Renamed feature directories
